heart: Unified: Refactor EROFS detection logic
Instead of checking potentially non-existing prop which could lead to wrong fstab parsing, Detect through super partition metadata instead.

This is still not the best approach but it's more reliable than the previous method and will do until we port linux version of blkid to android as an alternative.

Additionally cleanup the script for some unnecessary echos for new lines in fstab.

Change-Id: I2e597421eea846282c7db500b31bebf74095ae0a
Signed-off-by: ZorEl212 <yabsirad212@proton.me>